+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ## **WORK IN PROGRESS**
 - feature: add txtInsertButtons option to string fieldtype [documentation](https://book.quiqr.org/docs/20-quiqr-developer-reference/03-content-model/03-form-fields/data-field-types/string/)
 - feature: add loadingCircle when actionButton is busy
+- feature: add loadingCircle when AI popup is busy
+- feature: add buildActions to single form
 
 ## 0.19.2 (2025-01-30)
 - fix: missing build_actions cause form to crash